_quarantine/ADBMsysteam_deploy/get_data/requirements.txt
requests>=2.31.0
aiohttp>=3.9.0
websocket-client>=1.7.0
python-dotenv>=1.0.0
google-api-python-client>=2.151.0
google-auth>=2.34.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.2.1
redis>=5.0.0
pandas>=2.0.0
streamlit>=1.28.0
plotly>=5.0.0
pytz>=2024.1
psutil>=5.9.0

# US market holidays calendar
holidays>=0.89
